% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CPICGuidelines}
\alias{CPICGuidelines}
\title{Genes covered by CPIC Guidelines}
\format{a dataframe with 14 observations and one variable
\describe{
\item{gene}{gene name}
}}
\usage{
CPICGuidelines
}
\description{
Genes covered by CPIC Guidelines
}
\examples{
View(CPICGuidelines)
}
\keyword{datasets}
