% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitting.R
\name{fit_cl}
\alias{fit_cl}
\title{Fit a dose-response model for a single cell line}
\usage{
fit_cl(data, cell_line_name, model = drc::LL.4())
}
\arguments{
\item{data}{Aggregated data frame from \code{\link[=combine_reps]{combine_reps()}}.}

\item{cell_line_name}{Character. Name of the cell line to fit.}

\item{model}{A \code{drc} model function. Default \code{drc::LL.4()}. Other options
include \code{drc::LL.3()} (fixed lower asymptote at 0) or \code{drc::W1.4()}.}
}
\value{
A \code{drc} model object.
}
\description{
Wraps \code{\link[drc:drm]{drc::drm()}} for a single cell line subset of the aggregated data.
The default 4-parameter log-logistic model (\code{LL.4}) is appropriate for
sigmoidal dose-response relationships.
}
\examples{
\dontrun{
agg  <- combine_reps(load_data("data.csv"))
fit  <- fit_cl(agg, "SKMM-1")
}
}
