% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgxr-package.R
\docType{package}
\name{pgxr-package}
\alias{pgxr}
\alias{pgxr-package}
\title{pgxr: Creates a Pipeline for PGx Panel Analysis in R}
\description{
The goal of `pgxr` is to provide tools to create a pipeline for 
    pharmacogenomic panel analysis in r. It currently is based off of published 
    guidelines available at [CPIC](https://cpicpgx.org/guidelines/). Future work
    will include PharmGKB information, as well as other guidelines.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/rdsx50/pgxr}
  \item Report bugs at \url{https://github.com/rdsx50/pgxr/issues}
}

}
\author{
\strong{Maintainer}: Zach Rivers \email{river317@umn.edu}

}
\keyword{internal}
