% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{standardize_ancestry}
\alias{standardize_ancestry}
\title{Map ancestry labels to HGDP+1KP super pops}
\usage{
standardize_ancestry(x)
}
\arguments{
\item{x}{Character vectors of ancestry label}
}
\value{
Character vector with standardized abbreviations
}
\description{
Helpfer function for users whose data uses full pop names rather than standard abbrevs.
Unrecognized labels are passed through unchanged with a warning.
}
\examples{
standardize_ancestry(c("African", "European", "East Asian"))
}
