% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{validate_columns}
\alias{validate_columns}
\title{Validate required columns in a data frame}
\usage{
validate_columns(data, required, arg_name = "data")
}
\arguments{
\item{data}{A data frame.}

\item{required}{Character vector of required column names.}

\item{arg_name}{Name of the argument being checked, used in error messages.}
}
\description{
Validate required columns in a data frame
}
\keyword{internal}
