% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\docType{data}
\name{ANCESTRY_COLORS}
\alias{ANCESTRY_COLORS}
\title{Default ancestry color palette}
\format{
An object of class \code{character} of length 5.
}
\usage{
ANCESTRY_COLORS
}
\description{
A named character vector mapping HGDP/1KG superpopulation abbrevs to hex colors.
Pass a custom named vector to any plot function's
\code{ancestry_colors} argument to override.
}
\details{
Superpopulastions follow the HGDP + 1KGP callset standard:
\itemize{
\item AFR - African
\item AMR - Admixed American
\item EAS - East Asian
\item EUR - European
\item SAS - South Asian
}
}
\keyword{datasets}
